Added: Shortname & Fixed: Flame Predator's String Message
Added the new .ini attribute (shortname) to all of the standard races. This will make the races' name appear in the clan tag section if you've enabled this option in your wcs config file.

An issue with the Flame Predator's multi-language string was also solved, so it'll display the correct and intended information.